Update deps and linters (#5)
* Update dependencies
* Replace flake8 by ruff
* ruff: Add rules: flake8-builtins
* ruff: Add rules: flake8-blind-except
* ruff: Add rules: flake8-comprehensions
* ruff: Add rules: flake8-eradicate
* ruff: Add rules: mccabe
* ruff: Add rules: flake8-logging-format
* ruff: Add rules: flake8-isort
* ruff: Add rules: pep8-naming
* ruff: Add rules: flake8-pytest-style
* ruff: Add rules: ruff-specific-rules
* ruff: Add rules: flake8-bandit

Drive by:
* fix: Fix crash when listing caches if cache dir not present
* Do a full checkout in the action to fix issues during integration tests

+++ pipeline_runner/cli.py
@@ -175,10 +175,13 @@ def parse(pipeline, repository_path):
 @click.argument("action", type=click.Choice(["clear", "list"]))
 def cache(action):
     cache_dir = utils.get_cache_directory()
+    if not os.path.isdir(cache_dir):
+        return
+
     projects = sorted(os.listdir(cache_dir))
     if action == "list":
         print("Caches:")
-        print("\n".join(map(lambda i: f"\t{i}", projects)))
+        print("\n".join(f"\t{p}" for p in projects))
     elif action == "clear":
         for p in projects:
             shutil.rmtree(os.path.join(cache_dir, p))
